Present standard files page if the project is private, otherwise use the elasticsearch file page.
